// ws-client/src/app/school-erp/common/shared/model/education/education-session-institute.ts
export class EducationSessionInstitute {
  oid: String;
  nameEn: String;
  nameBn: String;
  status: String;
  enrolled: boolean;
  noOfInstitute: Number;
  noOfSessionEnrolledInstitute: Number;
  educationSystemOid: String;
  educationCurriculumOid: String;


  instituteNameEn: String;
  instituteNameBn: String;
  educationSystemNameEn: String;
  educationSystemNameBn: String;
  educationCurriculumNameEn: String;
  educationCurriculumNameBn: String;
}